Added support for % positioning of cover screens
Also inverted the values so that HA positioning system is reported correctly with 0% = closed and 100% = open.
This change implies that for some cover/shade systems it is necessary to change the runtime setting in the xcomfort app for each cover/shade from Manual to Automatic and run a single calibration.